Remove the duplicates removal code.
I'm a bit skeptical about that now: if we would introduce 'linear' option and
let statistics to be generated with reduced number of combinations the
burden of duplicated indexes would be no less than caused by it additional
statistics.

Also, tests with postfix '_0' added for patched core with the 'options' field
introduced.